% Auto-generated: do not edit by hand
\name{reactTableDash}

\alias{reactTableDash}

\title{ReactTableDash component}

\description{
ReactTableDash is a wrapper for ReactTable v6
}

\usage{
reactTableDash(id=NULL, data=NULL, columns=NULL, className=NULL, defaultPageSize=NULL,
style=NULL, showPagination=NULL, showPaginationTop=NULL,
showPaginationBottom=NULL, showPageSizeOptions=NULL, pageSizeOptions=NULL,
filterable=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{data}{Unnamed list. Table data}

\item{columns}{Unnamed list. Table columns}

\item{className}{Character. Table class name}

\item{defaultPageSize}{Numeric. Default page size}

\item{style}{Named list. Styles}

\item{showPagination}{Logical. Showing pagination (bool)}

\item{showPaginationTop}{Logical. Showing pagination on top (bool)}

\item{showPaginationBottom}{Logical. Showing pagination at bottom (bool)}

\item{showPageSizeOptions}{Logical. Page size options (list of integers)}

\item{pageSizeOptions}{Unnamed list. Page size options}

\item{filterable}{Logical. Table is filterable (true/false)}
}
